Revert "navbar added in app.js, css updated-audience/participant ranking/userProfile"
