+++ src/AllRoutes.jsx
@@ -9,6 +9,7 @@ import UserPortfolio from './Pages/UserPortfolio/UserPortfolio'
 import Portfolio from './Pages/Portfolios/Portfolio'
 import UserProfile from './Pages/UserProfile/UserProfile'
 import AudienceRanking from './Pages/AudienceRanking/AudienceRanking'
+import { IoMdPower } from 'react-icons/io'
 
 const AllRoutes = () => {
   return (
