+++ src/Pages/AudienceRanking/AudienceRanking.css
@@ -3,7 +3,6 @@
 .audience-ranking {
     width: 100%;
     min-height: 100vh;
-    padding: 72px 0 0 0;
 }
 
 .audience-ranking > h1 {
