+++ src/Pages/BuyPage/BuyPage.css
@@ -22,8 +22,7 @@
     -moz-border-radius: 10px;
     -ms-border-radius: 10px;
     -o-border-radius: 10px;
-    margin-top: 126px;
-    margin-bottom: 6px;
+    margin-top: 70px;
     align-items: center;
 }
