+++ src/Pages/Navbar/Navbar.css
@@ -15,13 +15,13 @@
     font-size: 36px;
 }
 
+
 /********Instruction Modal Window ******/
 
 .modal-overlay {
     background-color: #2c224e;
     color: #fff;
     padding: 28px 24px 10px 36px;
-
     /* background-image: linear-gradient(-45deg, #8b65eb 5%, #160f2dd1 70%); */
 }
 
@@ -39,11 +39,12 @@
 }
 
 .modal-overlay {
-    position: fixed;
+    position: absolute;
     filter:blur(0px);
-    width: 85%; 
+    width: 85%;
     left: 50%;
     top: 20px;
+
     z-index: 999;
     transform: translate(-50% ,150px);
     border-radius: 20px;
@@ -130,7 +131,6 @@ body {
     top: 0;
     right: 0;
     border-radius: 10px;
-    border: none;
     position: absolute;
     top: 86px;
     right: 14px;
