--- src/Pages/ParticipantRanking/ParticipantRanking.css
@@ -1,3 +1,0 @@
-.participants-rank {
-    padding-top: 75px;
-}
