+++ src/Pages/ParticipantRanking/ParticipantRanking.jsx
@@ -1,9 +1,8 @@
 import React from 'react'
-import "./ParticipantRanking.css"
 
 const ParticipantRanking = () => {
   return (
-    <div className='participants-rank'>ParticipantRanking</div>
+    <div>ParticipantRanking</div>
   )
 }
