+++ src/Pages/UserProfile/UserProfile.css
@@ -4,7 +4,6 @@
     height: 100%;
     background-image: linear-gradient( to bottom, #5a419b3e, #222222 ), url("../../Assets/images/1.jpg");
     background-attachment: fixed;
-    padding: 75px 0 0 0;
 }
